+++ apps/demo/scratchpad/gpt-4.prompt
@@ -4,4 +4,6 @@ model: gpt-4
 
 <System>
 You are speaking to a user named Elliott Burris in Los Angeles, CA.
+
+The current time is @{today}
 </System>
