+++ packages/rig/src/SessionView.tsx
@@ -1,0 +1,214 @@
+import { VSCodeTextArea } from '@vscode/webview-ui-toolkit/react'
+import { useEffect, useRef, useState } from 'react'
+import ReactMarkdown from 'react-markdown'
+import rehypeMathjax from 'rehype-mathjax'
+import remarkGfm from 'remark-gfm'
+import remarkMath from 'remark-math'
+import { CodeBlock } from './CodeBlock'
+import { ComposerView } from './ComposerView'
+import { ChatBlock } from './rig'
+
+interface SessionViewProps {
+  blocks: ChatBlock[]
+  theme: string
+  runChat: (chatToRun: string) => void
+  stop: () => void
+  streaming: boolean
+  session: string
+  inputs: Record<string, string>
+  setValue: (key: string, value: string) => void
+}
+
+export const SessionView = (props: SessionViewProps) => {
+  const { blocks, streaming, runChat, stop, theme, session, inputs, setValue } = props
+  const sessionId = session.split('/').pop()
+  const [autoScroll, setAutoScroll] = useState(true)
+  const chatContainer = useRef<HTMLDivElement | null>(null)
+
+  const capitalize = (s: string) => {
+    if (typeof s !== 'string') return ''
+    return s.charAt(0).toUpperCase() + s.slice(1)
+  }
+
+  const colorLookup: Record<string, string> = {
+    user: '#5EC5E5',
+    assistant: '#4EC9B0',
+    system: '#f467c6',
+    function: '#3EA295',
+  }
+
+  const handleScroll = () => {
+    if (!chatContainer.current) return
+
+    const { scrollTop, scrollHeight, clientHeight } = chatContainer.current
+    const atBottom = Math.abs(scrollHeight - scrollTop - clientHeight) < 5
+
+    setAutoScroll(atBottom)
+  }
+
+  useEffect(() => {
+    // Attach the scroll event handler
+    const current = chatContainer.current
+    chatContainer.current?.addEventListener('scroll', handleScroll)
+
+    // Detach the handler when the component unmounts
+    return () => {
+      current?.removeEventListener('scroll', handleScroll)
+    }
+  }, [])
+
+  useEffect(() => {
+    // Only scroll automatically if the user is at the bottom of the chat
+    if (autoScroll) {
+      document.getElementById('end')?.scrollIntoView()
+    }
+  }, [blocks, autoScroll])
+
+  return (
+    <div style={{ width: '100%', height: '100%', display: 'flex', flexDirection: 'column', overflow: 'hidden' }}>
+      {Object.keys(inputs).length > 0 ? (
+        <div
+          style={{
+            flexDirection: 'column',
+            overflowY: 'auto',
+            overflowX: 'hidden',
+            height: '100%',
+            paddingLeft: '24px',
+            paddingRight: '24px',
+          }}
+        >
+          {Object.keys(inputs).map(key => (
+            <div key={key} style={{ display: 'flex', flexDirection: 'column', paddingTop: '48px' }}>
+              <div style={{ paddingBottom: '4px' }}>{key}</div>
+              <VSCodeTextArea value={inputs[key]} onChange={(e: any) => setValue(key, e.target.value)} />
+            </div>
+          ))}
+        </div>
+      ) : (
+        <div
+          ref={chatContainer}
+          style={{
+            flexDirection: 'column',
+            overflowY: 'auto',
+            overflowX: 'hidden',
+            height: '100%',
+            paddingLeft: '24px',
+            paddingRight: '24px',
+          }}
+        >
+          <div
+            style={{
+              paddingTop: '16px',
+              paddingBottom: '16px',
+              fontFamily: 'monospace',
+              fontSize: '10px',
+              opacity: 0.3,
+              width: '100%',
+              textAlign: 'center',
+            }}
+          >
+            {sessionId}
+          </div>
+          {blocks
+            .filter(b => b.content !== `@{input}`)
+            .map((block, index) => {
+              return (
+                <div
+                  key={index}
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'column',
+                    paddingBottom: '48px',
+                    fontStyle: block.role === 'system' ? 'italic' : 'normal',
+                    width: '100%',
+                  }}
+                >
+                  <div style={{ display: 'flex', width: '100%', justifyContent: 'space-between' }}>
+                    <span
+                      style={{
+                        fontWeight: 'bold',
+                        color: block.role ? colorLookup[block.role] : undefined,
+                        fontStyle: 'italic',
+                        fontSize: '12px',
+                      }}
+                    >
+                      {capitalize(block.role)}
+                    </span>
+                    {/* {model && <span style={{ fontFamily: 'monospace', opacity: 0.5, fontSize: '10px' }}>{model}</span>} */}
+                  </div>
+                  <ReactMarkdown
+                    remarkPlugins={[remarkGfm, remarkMath]}
+                    rehypePlugins={[rehypeMathjax]}
+                    components={{
+                      code({ node, inline, className, children, ...props }) {
+                        const match = /language-(\w+)/.exec(className || '')
+                        return !inline && match ? (
+                          <CodeBlock language={match[1]} value={String(children).replace(/\n$/, '')} {...props} />
+                        ) : (
+                          <code className={className} {...props}>
+                            {children}
+                          </code>
+                        )
+                      },
+                      table({ children }) {
+                        return (
+                          <table
+                            style={{
+                              borderCollapse: 'collapse',
+                              border: '1px solid black',
+                              padding: '1rem 0.75rem',
+                              // Consider using a dark mode library to handle dark mode styles
+                            }}
+                          >
+                            {children}
+                          </table>
+                        )
+                      },
+                      th({ children }) {
+                        return (
+                          <th
+                            style={{
+                              // wordBreak: 'break-word',
+                              border: '1px solid black',
+                              backgroundColor: 'gray',
+                              padding: '1rem 0.75rem',
+                              color: 'white',
+                              // Consider using a dark mode library to handle dark mode styles
+                            }}
+                          >
+                            {children}
+                          </th>
+                        )
+                      },
+                      td({ children }) {
+                        return (
+                          <td
+                            style={{
+                              // wordBreak: 'break-word',
+                              border: '1px solid black',
+                              padding: '1rem 0.75rem',
+                              // Consider using a dark mode library to handle dark mode styles
+                            }}
+                          >
+                            {children}
+                          </td>
+                        )
+                      },
+                    }}
+                  >
+                    {block.content}
+                  </ReactMarkdown>
+                  {/* <div style={{ fontSize: '10px', opacity: 0.3, display: 'flex', flexDirection: 'column' }}>
+                {request && <span>{request}</span>}
+                {response && <span>{response}</span>}
+              </div> */}
+                </div>
+              )
+            })}
+          <div id={'end'} style={{ width: '100%', height: '0px' }} />
+        </div>
+      )}
+      <ComposerView theme={theme} runChat={runChat} stop={stop} streaming={streaming} />
+    </div>
+  )
+}
